feat: gorm datasource delete and upate

+++ linker/filter/filter.go
@@ -1,7 +1,7 @@
 package filter
 
 type Filter struct {
-	Skip  *int                    `json:"skip"`
-	Limit *int                    `json:"limit"`
+	Skip  *int64                  `json:"skip"`
+	Limit *int64                  `json:"limit"`
 	Where *map[string]interface{} `json:"where"`
 }
